# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Geenbank Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dossiers/{dossierId}/submissions:
    get:
      operationId: listDossierSubmissions
      tags: [submissions]
      summary: List partner submissions for a dossier
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PartnerSubmission" }
    post:
      operationId: submitDossierToPartners
      tags: [submissions]
      summary: Send dossier package to selected partner financiers
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerSubmissionRequest" }
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PartnerSubmission" }
...
components:
  schemas:
...
    PartnerSubmission:
      type: object
      properties:
        id: { type: string }
        dossierId: { type: string }
        partnerId: { type: string }
        partnerName: { type: string }
        status: { type: string }
        submittedAt: { type: string, nullable: true }
        packageSummary: { type: string, nullable: true }
        responseStatus: { type: string, nullable: true }
        responseNotes: { type: string, nullable: true }
        usedMockMode: { type: boolean }
      required: [id, dossierId, partnerId, partnerName, status, usedMockMode]
    PartnerSubmissionRequest:
      type: object
      properties:
        partnerIds:
          type: array
          items: { type: string }
        notes: { type: string, nullable: true }
      required: [partnerIds]
...